requirements.txt
setuptools
torch
transformers
pyahocorasick
pandas
datasets
gradio
tqdm